reporte de inventario casi completo

+++ app/Http/Controllers/IccController.php
@@ -21,7 +21,6 @@ class IccController extends Controller
      */
     public function index()
     {
-
     }
 
     /**
@@ -43,125 +42,125 @@ class IccController extends Controller
     public function store(Request $request)
     {
         $user = Auth::user();
-        
-        if($user->can('store stock')){
-       //array de mensajes de error y exitosos
-        $errores = [];
 
-        $exitosos = [];
+        if ($user->can('store stock')) {
+            //array de mensajes de error y exitosos
+            $errores = [];
 
-        $series = json_decode($request->data);
+            $exitosos = [];
 
-        $inventario = $request->input('inventario_id');
+            $series = json_decode($request->data);
 
-        $company = $request->input('company_id');
+            $inventario = $request->input('inventario_id');
 
-        $type = $request->input('icc_type_id');
+            $company = $request->input('company_id');
 
-
-        //si el request contiene un archivo excel procede con esta funcion 
-        if ($request->hasFile('file')) {
-
-            $file = $request->file('file');
+            $type = $request->input('icc_type_id');
 
 
+            //si el request contiene un archivo excel procede con esta funcion 
+            if ($request->hasFile('file')) {
 
-            //datos enviados al import que vienen desde la request
-            $data = [
-                'inventario_id' => $inventario,
-                'status_id' => 1,
-                'icc_type_id' => $type,
-                'company_id' => $company,
-
-            ];
-
-
-            $imeiImport = new IccsImport($data);
-            $imeiImport->import($file);
-
-            //obtiene los los mensales de error
-            $imeiValidationErrors = $imeiImport->getErrors();
-            $imeiValidationSuccess = $imeiImport->getsuccess();
-
-            //pushea los mensajes a los arrays
-            foreach ($imeiValidationErrors as $validationErr) {
-                $errors = array_push($errores, $validationErr);
-            }
-            foreach ($imeiValidationSuccess as $validationSucc) {
-                $errors = array_push($exitosos, $validationSucc);
-            }
-        }
+                $file = $request->file('file');
 
 
 
-        if ($series) {
-
-            foreach ($series as $datos) {
-
-                $serie = [];
-
-                $serie = ['serie' => $datos->serie];
-
-                
-
-                $icc = new Icc([
-                    
-                    'icc' => $datos->serie,
-                    'status_id' => 1,
+                //datos enviados al import que vienen desde la request
+                $data = [
                     'inventario_id' => $inventario,
-                    'company_id' => $company,
+                    'status_id' => 1,
                     'icc_type_id' => $type,
-                   
+                    'company_id' => $company,
+
+                ];
 
 
-                ]);
+                $imeiImport = new IccsImport($data);
+                $imeiImport->import($file);
 
-                // Crea la matriz de mensajes.
-                $mensajes = array(
-                    'unique' => 'ya existe en la base de datos.',
-                    'digits' => 'La serie tiene que se numerica y de 19 digitos'
-                );
+                //obtiene los los mensales de error
+                $imeiValidationErrors = $imeiImport->getErrors();
+                $imeiValidationSuccess = $imeiImport->getsuccess();
 
-
-                $validator = Validator::make($serie, [
-                    'serie' => 'required|unique:iccs,icc|digits:19',
-
-
-
-                ], $mensajes);
-                if ($validator->fails()) {
-
-                    $err = [];
-
-                    $errorList = [];
-
-                    $err['serie'] = $datos->serie;
-
-
-
-                    foreach ($validator->errors()->toArray() as $error) {
-
-
-
-                        foreach ($error as $sub_error) {
-                            array_push($errorList, $sub_error);
-                        }
-                    }
-                    $err['errores'] = $errorList;
-                    array_push($errores, $err);
-                } else {
-
-                    $icc->save();
-                    $icc->setStatus('Disponible');
-                    array_push($exitosos, $serie);
+                //pushea los mensajes a los arrays
+                foreach ($imeiValidationErrors as $validationErr) {
+                    $errors = array_push($errores, $validationErr);
+                }
+                foreach ($imeiValidationSuccess as $validationSucc) {
+                    $errors = array_push($exitosos, $validationSucc);
                 }
             }
-        }
 
-        //regresa los mensajes de errores y exitosos
-        return ['errors' => $errores, 'success' => $exitosos];
-    }
-    return ['errors' => ['usuario sin permisos']];
+
+
+            if ($series) {
+
+                foreach ($series as $datos) {
+
+                    $serie = [];
+
+                    $serie = ['serie' => $datos->serie];
+
+
+
+                    $icc = new Icc([
+
+                        'icc' => $datos->serie,
+                        'status_id' => 1,
+                        'inventario_id' => $inventario,
+                        'company_id' => $company,
+                        'icc_type_id' => $type,
+
+
+
+                    ]);
+
+                    // Crea la matriz de mensajes.
+                    $mensajes = array(
+                        'unique' => 'ya existe en la base de datos.',
+                        'digits' => 'La serie tiene que se numerica y de 19 digitos'
+                    );
+
+
+                    $validator = Validator::make($serie, [
+                        'serie' => 'required|unique:iccs,icc|digits:19',
+
+
+
+                    ], $mensajes);
+                    if ($validator->fails()) {
+
+                        $err = [];
+
+                        $errorList = [];
+
+                        $err['serie'] = $datos->serie;
+
+
+
+                        foreach ($validator->errors()->toArray() as $error) {
+
+
+
+                            foreach ($error as $sub_error) {
+                                array_push($errorList, $sub_error);
+                            }
+                        }
+                        $err['errores'] = $errorList;
+                        array_push($errores, $err);
+                    } else {
+
+                        $icc->save();
+                        $icc->setStatus('Disponible');
+                        array_push($exitosos, $serie);
+                    }
+                }
+            }
+
+            //regresa los mensajes de errores y exitosos
+            return ['errors' => $errores, 'success' => $exitosos];
+        }
+        return ['errors' => ['usuario sin permisos']];
     }
 
     /**
@@ -196,22 +195,70 @@ class IccController extends Controller
     public function update(Request $request, $id)
     {
         $user = Auth::user();
-        if($user->can('full update stock')){
-        
+
         $icc = Icc::findorfail($id);
 
+        if ($user->can('update stock')) {
 
-        $icc->update($request->all());
+            if ($request->comment != NULL) {
 
+                $icc->comment()->updateOrCreate([], ['comment' => $request->comment]);
+            } else {
 
-        if ($request->comment != NULL) {
+                $icc->comment()->delete();
+            }
+            $icc->setStatus($request->status);
+        
+            if ($user->can('full update stock')) {
 
-            $icc->comment()->updateOrCreate([], ['comment' => $request->comment]);
-        } else {
+                $icc->inventario_id = $request->inventario_id;
 
-            $icc->comment()->delete();
+                $icc->company_id = $request->company_id;
+
+                $icc->icc_type_id = $request->icc_type_id;
+
+                if($request->lineaDn){
+                    
+                    $linea = $icc->linea;
+
+                    $linea->dn = $request->lineaDn;
+
+                    $montoRecarga= null;
+
+                    $lineaStatus = null;
+
+                    if($request->lineaStatus == 'Recargable'){
+                        if($user->hasPermissionTo('asignar recarga')){
+                            $montoRecarga = $request->montoRecarga;
+
+                            $lineaStatus = $request->lineaStatus;
+
+                            $linea->setStatus($lineaStatus,$montoRecarga);
+                        }
+
+                    }else{
+                        $lineaStatus = $request->lineaStatus;
+
+                        $montoRecarga = null;
+
+                        $linea->setStatus($lineaStatus,$montoRecarga);
+                    }
+
+                    
+
+                    
+
+                    $linea->save();
+                }
+
+                
+
+            }
+
+            $icc->save();
+            
         }
-    }
+        return $request;
     }
 
     /**
@@ -223,10 +270,38 @@ class IccController extends Controller
     public function destroy(Icc $icc)
     {
         $user = Auth::user();
-        if($user->can('destroy stock')){
-        $icc->details()->delete();
+        if ($user->can('destroy stock')) {
 
-        $icc->delete();
+            $linea = $icc->linea();
+
+            $linea->delete();
+
+            $icc->setStatus('Eliminado');
+
+            $icc->delete();
         }
+    }
+
+
+    public function restore(Request $request)
+    {
+        $user = Auth::user();
+        if ($user->can('destroy stock')) {
+        $id = $request;
+        
+        $icc = Icc::onlyTrashed()->findOrfail($id)->first();
+
+        $icc->restore();
+
+        $icc->deleteStatus('Eliminado');
+
+        $linea = $icc->linea()->withTrashed();
+
+        $linea->restore();
+
+        return $icc;
+        }
+
+        
     }
 }
